INS-2263: refactor genesis bootstrap
